refactor(web): extract dockview layout hooks
